backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
requests
python-multipart
httpx